Move SettingsMenu out of App
